Secure routes

+++ api/film.py
@@ -6,12 +6,14 @@ from internal.models.film import FilmBase, FilmCreate
 from internal.repository.film_repository import FilmRepository
 from starlette.responses import JSONResponse
 
+from api.additions.additions import auth_required
+
 router = APIRouter(prefix="/films", tags=['films'])
 
 oauth2_scheme = OAuth2PasswordBearer(tokenUrl="api/users", auto_error=True)
 
 
-@router.get("", response_model=list[FilmBase])
+@router.get("", response_model=list[FilmBase],)
 async def film_list(film_repo: FilmRepository = Depends(get_repository(FilmRepository))) -> list[FilmBase]: # noqa
     film_list = await film_repo.get()
     return film_list
@@ -20,14 +22,13 @@ async def film_list(film_repo: FilmRepository = Depends(get_repository(FilmRepos
 @router.get("/{id}",
             response_model=FilmBase)
 async def film_one(id: int,
-                   film_repo: FilmRepository = Depends(get_repository(FilmRepository)),
-                   security: OAuth2PasswordBearer=Depends(oauth2_scheme)) -> FilmBase: # noqa
+                   film_repo: FilmRepository = Depends(get_repository(FilmRepository))): # noqa
     fetched_film = await film_repo.get_one(id)
     if fetched_film is None:
         return JSONResponse(
-                status_code=status.HTTP_404_NOT_FOUND,
-                content={"message": "not found"}
-            )
+            status_code=status.HTTP_404_NOT_FOUND,
+            content={"message": "not found"}
+        )
 
     return fetched_film
 
@@ -35,26 +36,38 @@ async def film_one(id: int,
 @router.post("",
              status_code=status.HTTP_201_CREATED)
 async def create_film(new_film: FilmCreate,
-                      film_repo: FilmRepository = Depends(get_repository(FilmRepository))): # noqa
+                      film_repo: FilmRepository = Depends(get_repository(FilmRepository)), # noqa
+                      authorized: bool = Depends(auth_required)):
+    if not authorized:
+        return JSONResponse(
+            status_code=401,
+            content={"message": "no auth"}
+        )
     try:
         await film_repo.create(new_film)
         film_repo.logger.info(f"Создан фильм под названием: {new_film.title}")
     except UniqueViolationError:
         film_repo.logger.info(f"Фильм под названием {new_film.title} уже существует") # noqa
         return JSONResponse(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                content={"message": "Film already exists"}
-            )
+            status_code=status.HTTP_400_BAD_REQUEST,
+            content={"message": "Film already exists"}
+        )
 
     return JSONResponse(
-            status_code=status.HTTP_201_CREATED,
-            content={"success": "created"}
-        )
+        status_code=status.HTTP_201_CREATED,
+        content={"success": "created"}
+    )
 
 
 @router.delete("/{id}")
 async def delete_film(id: int,
-                      film_repo: FilmRepository = Depends(get_repository(FilmRepository))): # noqa
+                      film_repo: FilmRepository = Depends(get_repository(FilmRepository)), # noqa
+                      authorized: bool = Depends(auth_required)):
+    if not authorized:
+        return JSONResponse(
+            status_code=401,
+            content={"message": "no auth"}
+        )
     try:
         deleted = await film_repo.delete(id)
     except Exception as e:
@@ -78,8 +91,13 @@ async def delete_film(id: int,
 @router.put("/{id}")
 async def update_film(id: int,
                       updated_film: FilmCreate,
-                      film_repo: FilmRepository = Depends(get_repository(FilmRepository))): # noqa
-
+                      film_repo: FilmRepository = Depends(get_repository(FilmRepository)), # noqa
+                      authorized: bool = Depends(auth_required)):
+    if not authorized:
+        return JSONResponse(
+                status_code=401,
+                content={"message": "no auth"}
+            )
     updated_film = FilmBase(id=id, **updated_film.dict())
     try:
         updated = await film_repo.update(updated_film)
@@ -96,6 +114,6 @@ async def update_film(id: int,
         )
 
     return JSONResponse(
-            status_code=status.HTTP_200_OK,
-            content={"message": f"Updated:{id}"}
-        )
+        status_code=status.HTTP_200_OK,
+        content={"message": f"Updated:{id}"}
+    )
